add underline detect image

+++ quicktools/pane/custom_checked_listview.h
@@ -9,7 +9,9 @@
 #define NORMAL_IMAGE_FILE L"normal.tiff"
 #define IMAGE_FILE_MINIMIZE_SIZE 32
 
+// 打开图像文件时替换当前活动视图
 #define OPEN_IMAGE_THIS 0
+// 在新视图中打开图像文件
 #define OPEN_IMAGE_NEW 1
 
 namespace freeze {
@@ -23,21 +25,86 @@ namespace freeze {
 		std::wstring loc;
 	};
 
-	// 带选择框的列表视图
-	// 顶部添加了筛选框
-	// 底部添加了目录框
+	class CMyCheckedListViewCtrl :
+		public WTL::CCheckListViewCtrlImpl<CMyCheckedListViewCtrl>
+	{
+	public:
+		DECLARE_WND_SUPERCLASS(_T("WTL_CheckListView"), GetWndClassName())
+
+		BEGIN_MSG_MAP_EX(CMyCheckedListViewCtrl)
+			CHAIN_MSG_MAP(WTL::CCheckListViewCtrlImpl<CMyCheckedListViewCtrl>)
+		END_MSG_MAP()
+
+		void CheckSelectedItems(int nCurrItem)
+		{
+			// 用户勾选/取消勾选了此选项
+			GetParent().PostMessage(WM_CHECKED_ITEM, nCurrItem, 0);
+
+			// first check if this item is selected
+			LVITEM lvi = {};
+			lvi.iItem = nCurrItem;
+			lvi.iSubItem = 0;
+			lvi.mask = LVIF_STATE;
+			lvi.stateMask = LVIS_SELECTED;
+			this->GetItem(&lvi);
+			// if item is not selected, don't do anything
+			if (!(lvi.state & LVIS_SELECTED))
+				return;
+			// new check state will be reverse of the current state,
+			BOOL bCheck = !this->GetCheckState(nCurrItem);
+			int nItem = -1;
+			int nOldItem = -1;
+			while ((nItem = this->GetNextItem(nOldItem, LVNI_SELECTED)) != -1)
+			{
+				if (nItem != nCurrItem)
+					this->SetCheckState(nItem, bCheck);
+				nOldItem = nItem;
+			}
+		}
+	};
+
+	//
+	// +---------------------+
+	// |       搜索栏         |
+	// +---------------------+
+	// |     自定义列表视图     |
+	// +---------------------+
+	// |      目录浏览栏       |
+	// +---------------------+
+	//
+	// 用于: CDockingContainer
+	//
 	class CCustomCheckedListview :
 		public ATL::CWindowImpl<CCustomCheckedListview>
 	{
 	public:
 		DECLARE_WND_SUPERCLASS(_T("WTL_CustomCheckedListview"), GetWndClassName())
 
+		//
+		// +---------------------+
+		// |       搜索栏         |
+		// +---------------------+
+		// |     自定义列表视图     |
+		// +---------------------+
+		// |      目录浏览栏       |
+		// +---------------------+
+		//
 		CSearchBar m_SearchBar;
-		WTL::CCheckListViewCtrl m_CheckListViewCtrl;
+		CMyCheckedListViewCtrl m_CheckListViewCtrl;
 		CDirBar m_DirBar;
+
+		// 所有原始(待处理)图像信息容器
 		std::vector<ListViewData> m_VecData;
+
+		// 图像文件目录
 		std::wstring m_CurrentImageDirectory;
+		// 工作中的图像文件
 		std::wstring m_CurrentImageFile;
+
+		// 检测图像文件目录
+		std::wstring m_CurrentDetectImageDirectory;
+		// 工作中的检测图像文件 与工作图像一一对应
+		std::wstring m_CurrentDetectImageFile;
 
 		//TODO: 是否有参考图像
 		bool m_HasRefImage = false;
@@ -125,9 +192,26 @@ namespace freeze {
 			return false;
 		}
 
-		void ResetListView(std::wstring const& dir)
+		bool SetSelectedItem(int item)
 		{
-			ReadImageFiles(dir);
+			if (m_CheckListViewCtrl.IsWindow())
+			{
+				return m_CheckListViewCtrl.SelectItem(item) == TRUE;
+			}
+			return false;
+		}
+
+		void ResetListView(std::wstring const& dir,int dir_type = RESET_LISTVIEW_DIR)
+		{
+			if (dir_type == RESET_LISTVIEW_DIR_DETECT)
+			{
+				m_CurrentDetectImageDirectory = dir;
+				return;
+			}
+
+			m_CurrentImageDirectory = dir;
+			ReadImageFiles(m_CurrentImageDirectory);
+
 			m_CheckListViewCtrl.DeleteAllItems();
 
 			if (!m_CheckListViewCtrl.IsGroupViewEnabled())
@@ -162,7 +246,7 @@ namespace freeze {
 				group.iGroupId = 1;
 				group.mask = LVGF_HEADER | LVGF_GROUPID;
 				auto success = m_CheckListViewCtrl.AddGroup(&group);
-				
+
 				for (auto i = 0; i < m_VecData.size() - 1; ++i)
 				{
 					InsertItem(i, 1, m_VecData[i]);
@@ -234,6 +318,7 @@ namespace freeze {
 			MSG_WM_CREATE(OnCreate)
 			MSG_WM_SIZE(OnSize)
 			MESSAGE_HANDLER_EX(WM_RESET_LISTVIEW, OnResetListView)
+			MESSAGE_HANDLER_EX(WM_CHECKED_ITEM, OnListViewItemChecked)
 			MSG_WM_NOTIFY(OnNotify)
 			MSG_WM_DROPFILES(OnDropFiles)
 			MSG_WM_COMMAND(OnCommand)
@@ -258,7 +343,7 @@ namespace freeze {
 				header.ModifyStyle(0, HDS_CHECKBOXES);
 				header.SetDlgCtrlID(ID_CHECKED_HEADER);
 				HDITEM hdi = {};
-				hdi.mask = HDI_WIDTH|HDI_FORMAT;
+				hdi.mask = HDI_WIDTH | HDI_FORMAT;
 				hdi.fmt = HDF_CHECKBOX;
 				hdi.cxy = 80;
 				header.AddItem(&hdi);
@@ -288,16 +373,31 @@ namespace freeze {
 
 		LRESULT OnResetListView(UINT uMsg, WPARAM wParam, LPARAM lParam)
 		{
-			auto size = static_cast<unsigned>(wParam);
+			auto size = HIWORD(wParam);
+			auto dir_type = LOWORD(wParam);
+
 			wchar_t* pDir = new wchar_t[size];
 			wmemcpy_s(pDir, size, reinterpret_cast<wchar_t*>(lParam), size);
-			m_CurrentImageDirectory = std::wstring{ pDir,size };
+			auto dir = std::wstring{ pDir,size };
 			delete[] pDir;
 
-			ResetListView(m_CurrentImageDirectory);
+			ResetListView(dir,dir_type);
 
 			return 0;
 		}
+
+		// 当用户勾选了某个项时
+		LRESULT OnListViewItemChecked(UINT /*uMsg*/, WPARAM wParam, LPARAM /*lParam*/)
+		{
+			SetSelectedItem(static_cast<int>(wParam));
+			NMHDR nmh = {};
+			nmh.hwndFrom = m_CheckListViewCtrl;
+			nmh.idFrom = ID_CHECKED_LISTVIEW;
+			nmh.code = NM_CLICK;
+			OnNotify(ID_CHECKED_LISTVIEW, &nmh);
+			return 0;
+		}
+
 
 		LRESULT OnNotify(int idCtrl, LPNMHDR pnmh)
 		{
@@ -306,29 +406,46 @@ namespace freeze {
 			{
 				if (NM_CLICK == pnmh->code || NM_RCLICK == pnmh->code)
 				{
-					auto index = GetSelectedIndex() + 1;
-					if (index > 0)
+					auto index = GetSelectedIndex();
+					if (index >= 0)
 					{
+						auto checked = IsChecked(index);
+						m_HasRefImage = checked;
+
 						auto find = std::find_if(std::begin(m_VecData), std::end(m_VecData), [&index](auto data) {
-							return data.index == index;
+							return data.index == (index + 1);
 						});
 
 						if (m_VecData.end() != find)
 						{
 							m_CurrentImageFile = m_CurrentImageDirectory + L"\\" + find->file;
 						}
+
+						if (m_HasRefImage && !m_CurrentDetectImageDirectory.empty())
+						{
+							auto detect_file = find->file;
+							// 总长度是37，返回的是36
+							auto pos = detect_file.find_last_of(L".tiff");
+							// 替换的起始位置，替换多长
+							detect_file.replace(pos-4, 5, L".bmp");
+							m_CurrentDetectImageFile = m_CurrentDetectImageDirectory + L"\\" + detect_file;
+						}
+						else
+						{
+							m_CurrentDetectImageFile = L"";
+						}
 					}
 					else
 					{
 						// 可能单击的是前面的复选框
 						// TODO: 重载CheckListViewCtrl
-						CPoint pt;
-						GetCursorPos(&pt);
-						m_CheckListViewCtrl.ScreenToClient(&pt);
-						LVHITTESTINFO hitTestInfo = {};
-						hitTestInfo.pt = pt;
-						m_CheckListViewCtrl.HitTest(&hitTestInfo);
-						int item = hitTestInfo.iItem;
+						//CPoint pt;
+						//GetCursorPos(&pt);
+						//m_CheckListViewCtrl.ScreenToClient(&pt);
+						//LVHITTESTINFO hitTestInfo = {};
+						//hitTestInfo.pt = pt;
+						//m_CheckListViewCtrl.HitTest(&hitTestInfo);
+						//int item = hitTestInfo.iItem;
 					}
 
 					// 弹出菜单
@@ -340,7 +457,7 @@ namespace freeze {
 						ClientToScreen(&pt);
 						TrackPopupMenuEx(pop,
 							TPM_NOANIMATION,
-							pt.x,pt.y,
+							pt.x, pt.y,
 							this->m_hWnd,
 							nullptr);
 					}
@@ -348,7 +465,12 @@ namespace freeze {
 					{
 						if (!m_CurrentImageFile.empty())
 						{
-							GetParent().SendMessage(WM_OPEN_IMAGE, OPEN_IMAGE_THIS, reinterpret_cast<LPARAM>(m_CurrentImageFile.c_str()));
+							auto docking_ontainer = GetParent();
+							if (docking_ontainer)
+							{
+								docking_ontainer.SendMessage(WM_OPEN_IMAGE_WITH_DETECT, OPEN_IMAGE_THIS, reinterpret_cast<LPARAM>(m_CurrentDetectImageFile.c_str()));
+								docking_ontainer.SendMessage(WM_OPEN_IMAGE, OPEN_IMAGE_THIS, reinterpret_cast<LPARAM>(m_CurrentImageFile.c_str()));
+							}
 						}
 					}
 				}
@@ -360,7 +482,7 @@ namespace freeze {
 			}
 
 			// 当选择列表视图标题栏中的全选框时
-			if (auto header = m_CheckListViewCtrl.GetHeader(); header.GetDlgCtrlID()==idCtrl)
+			if (auto header = m_CheckListViewCtrl.GetHeader(); header.GetDlgCtrlID() == idCtrl)
 			{
 				HDITEM hdi;
 				if (header.GetItem(0, &hdi))
@@ -390,9 +512,11 @@ namespace freeze {
 			case ID_CLV_MENU_SELECT_ONE: // 勾选
 				break;
 			case ID_CLV_MENU_OPEN_THIS: // 当前窗口打开
+				GetParent().SendMessage(WM_OPEN_IMAGE_WITH_DETECT, OPEN_IMAGE_THIS, reinterpret_cast<LPARAM>(m_CurrentDetectImageFile.c_str()));
 				GetParent().SendMessage(WM_OPEN_IMAGE, OPEN_IMAGE_THIS, reinterpret_cast<LPARAM>(m_CurrentImageFile.c_str()));
 				break;
 			case ID_CLV_MENU_OPEN_NEW: // 新建窗口打开
+				GetParent().SendMessage(WM_OPEN_IMAGE_WITH_DETECT, OPEN_IMAGE_NEW, reinterpret_cast<LPARAM>(m_CurrentDetectImageFile.c_str()));
 				GetParent().SendMessage(WM_OPEN_IMAGE, OPEN_IMAGE_NEW, reinterpret_cast<LPARAM>(m_CurrentImageFile.c_str()));
 				break;
 			}
